# openapi.yaml
openapi: 3.0.3
info:
  title: Whop SaaS Marketplace API
  description: |
    Production-grade marketplace API enabling multi-tenant SaaS, Stripe Connect payments,
    license management, and webhook event broadcasting.

    ## Authentication
    All endpoints require authentication via session token cookie (`session_token`).

    ## Multi-Tenancy
    All resources are scoped to companies. Users can access multiple companies with different roles.

    ## Rate Limiting
    - 100 requests per minute per IP
    - Returns `429 Too Many Requests` when exceeded

    ## Security
    - Tenant isolation via `company_id` filtering
    - Role-based access control (owner/admin/member)
    - Money stored in minor units (cents) to avoid floating-point errors
  version: 1.0.0
  contact:
    name: API Support
    email: api@whopsaas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.whopsaas.com
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: JWT session token (HTTP-only cookie)
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_abc123
        company_id:
          type: string
          example: biz_xyz789
        title:
          type: string
          example: Pro Plan
        description:
          type: string
          nullable: true
          example: Premium features with priority support
        image_url:
          type: string
          format: uri
          nullable: true
        price_amount:
          type: number
          format: decimal
          example: 49.99
        price_minor_units:
          type: integer
          example: 4999
          description: Price in cents (avoid floating-point errors)
        currency:
          type: string
          example: USD
        plan_type:
          type: string
          enum: [one_time, monthly, yearly]
        billing_period:
          type: integer
          nullable: true
          description: Billing period in days (30, 365, etc.)
        trial_days:
          type: integer
          example: 14
        active:
          type: boolean
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - company_id
        - title
        - price_amount
        - price_minor_units
        - currency
        - plan_type
...
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        image_url:
          type: string
          format: uri
          nullable: true
        price_amount:
          type: number
        price_minor_units:
          type: integer
        currency:
          type: string
        plan_type:
          type: string
          enum: [one_time, monthly, yearly]
        billing_period:
          type: integer
          nullable: true
        trial_days:
          type: integer
        active:
          type: boolean
        metadata:
          type: object
          nullable: true

    # ========================================================================
    # Company Schemas
    # ========================================================================
    Company:
      type: object
      properties:
        id:
          type: string
          example: biz_abc123
        type:
          type: string
          enum: [platform, sub_merchant]
        status:
          type: string
          enum: [active, pending_kyc, suspended, closed]
        title:
          type: string
        email:
          type: string
          format: email
        slug:
          type: string
          example: my-saas-company
        logo_url:
          type: string
          format: uri
          nullable: true
        website_url:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        platform_fee_percent:
          type: number
          format: decimal
          example: 5.0
        payout_minimum_amount:
          type: number
          format: decimal
          example: 50.0
        payout_frequency:
          type: string
          enum: [daily, weekly, monthly]
        stripe_account_id:
          type: string
          nullable: true
        stripe_onboarded:
          type: boolean
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            products:
              type: integer
            memberships:
              type: integer
            payments:
              type: integer
            users:
              type: integer
...
    Membership:
      type: object
      properties:
        id:
          type: string
          example: mem_abc123
        user_id:
          type: string
        product_id:
          type: string
        company_id:
          type: string
        status:
          type: string
          enum: [active, past_due, canceled, expired, trialing]
        stripe_subscription_id:
          type: string
          nullable: true
        current_period_start:
          type: string
          format: date-time
          nullable: true
        current_period_end:
          type: string
          format: date-time
          nullable: true
        cancel_at:
          type: string
          format: date-time
          nullable: true
        canceled_at:
          type: string
          format: date-time
          nullable: true
        trial_end:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
            avatar_url:
              type: string
        product:
          $ref: '#/components/schemas/Product'
        company:
          $ref: '#/components/schemas/Company'
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
security:
  - cookieAuth: []
paths:
...
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    put:
      tags:
        - Products
      summary: Update product
      description: Update product (requires admin/owner role)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
        '403':
          description: Forbidden

    delete:
      tags:
        - Products
      summary: Delete product
      description: Soft delete (sets active=false). Only owners can delete. Prevents deletion if active memberships exist.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    $ref: '#/components/schemas/Product'
        '400':
          description: Cannot delete (active memberships exist)
        '403':
          description: Forbidden (only owners can delete)

  # ==========================================================================
  # COMPANIES
  # ==========================================================================
...
  /api/memberships:
    get:
      tags:
        - Memberships
      summary: List memberships
      description: List memberships (scoped by company_id or user)
      parameters:
        - name: company_id
          in: query
          schema:
            type: string
          description: Filter by company (requires admin access)
        - name: user_id
          in: query
          schema:
            type: string
        - name: product_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, past_due, canceled, expired, trialing]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberships:
                    type: array
                    items:
                      $ref: '#/components/schemas/Membership'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Forbidden
...